// src/app/modules/academics/events/model/event-view-model.ts
import { AttachmentViewModel } from '../../student-admissions/model/attachment-view-model';
import { EventClassSectionsView } from './event-class-sections-view';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets EventClassSectionView
 */
export interface EventView { 
    /**
     * Gets or Sets AnnounceDate
     */
    announcedDate?: any;
    /**
     * Gets or Sets FromDate
     */
    fromDate?: any;
    /**
     * Gets or Sets ToDate
     */
    toDate?: any;
    /**
     * Gets or Sets EventTypeId
     */
    eventTypeId?: number;
    /**
     * Gets or Sets StatusId
     */
    statusId?: number;
    /**
     * Gets or Sets EventDescription
     */
    eventDescription?: string;
    /**
     * Gets or set
     */
    attachment?: AttachmentViewModel;
     /**
     * Gets or Sets IsPostToCalender
     */
    isPostToCalender?: number;
    
    /**
     * Gets or Sets SchoolCalendarView
     */
    eventClassSections?: Array<EventClassSectionsView>;
    eventid: string;
}